Drop the dependency on array
The array package was imported to provide NFData instances for the
Array type. However, this type is defined in base, not array, making
the dependency unncecessary. The Data.Array import is replaced with
GHC.Arr.

This change will also allow array to define NFData instances for other
types, because it can now depend on deepseq.